requirements.txt
streamlit==1.25.0
pandas==2.0.3
joblib==1.3.1
scikit-learn==1.3.0
numpy==1.25.1
Pillow==9.5.0